Created TextArea class, edited Button class. Organized classes structure. Created components dir with scss files.
