+++ src/scss/_mixins.scss
@@ -1,7 +1,13 @@
 @mixin commonBtnText {
   font-family: Roboto, sans-serif;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 15px;
-  line-height: 22px;
+  font-size: 18px;
+  line-height: 18px;
+}
+
+@mixin resetButtonStyles{
+  outline: none;
+  display: block;
+  font: inherit;
+  cursor: pointer;
+  border: none;
 }
