+++ src/scss/_variables.scss
@@ -1,3 +1,4 @@
-$blue: #3FA7F5;
-$white: #F9F9F9;
-
+$buttonColor: #3FA7F5;
+$titleText: #3FA7F5;
+$buttonTextColor: #FFFFFF;
+$crossColor: #585F73;
