+++ src/scss/components/_form.scss
@@ -1,0 +1,20 @@
+.form{
+  padding: 30px 40px 30px 40px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.form-input{
+  text-align: center;
+  font-size: 24px;
+  padding: 0.3em;
+  line-height: 24px;
+  display: block;
+  margin: 10px 0;
+  width: 100%;
+  background: #fff;
+  border-radius: 5px;
+  color: #585f73;
+}
