+++ src/scss/components/_modal.scss
@@ -17,7 +17,6 @@
   align-items: center;
   width: 850px;
   position: relative;
-  border-radius: 30px;
 }
 
 .cross{
@@ -26,9 +25,10 @@
   top: 10px;
   border: none;
   background: transparent;
+  color: $crossColor;
 }
 .modal-title{
-  color: $blue;
+  color: $titleText;
   @include commonBtnText;
   font-weight: bold;
   font-size: 42px;
