+++ src/scss/index.scss
@@ -1,5 +1,8 @@
 @import "reset";
 @import "variables";
-@import "utils";
 @import "mixins";
-@import "modalLogin";
+@import "common";
+@import "components/createElements";
+@import "components/modal";
+@import "components/form";
+@import "components/visit";
